Moves liquid unit testing to it's own file so that it's not tested alongside browser tests
